our executable file
because of a minor bug in sound manager, the project was working when run from the netbeans but giving bugs when run from the jar. Until detecting the bug we missed the deadline about 30 minutes, now sound from the game is excluded but can be shown in the demo it's allowed to run the game from netbeans.